requirements.txt
apscheduler==3.10.4
fastapi==0.116.0
itsdangerous==2.2.0
jinja2==3.1.6
passlib[argon2]==1.7.4
pydantic-settings==2.11.0
python-multipart==0.0.20
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0
